Moved generic tests from Test_Admin_Menu to Test_Base_Admin_Menu
